backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.2
xlrd>=2.0
python-dotenv>=1.0
pydantic>=2.7
pydantic-settings>=2.3
weasyprint>=62
python-multipart>=0.0.9
httpx>=0.27
